+++ src/test/java/e2e/EarProjectTest.java
@@ -10,7 +10,6 @@ import static scaffolding.GitMatchers.hasTagWithModuleVersion;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -21,13 +20,9 @@ public class EarProjectTest {
     @Rule
     public TestProject testProject = new TestProject(ProjectType.EAR);
 
-    @Before
-    public void setUp() {
-        testProject.mvnReleaseComplete();
-    }
-
     @Test
     public void buildsAndInstallsAndTagsAllModules() throws Exception {
+        testProject.mvnReleaseComplete();
         assertThat(testProject.local, hasTagWithModuleVersion(TestUtils.TEST_GROUP_ID, "parent-project", "1.0"));
         assertThat(testProject.local, hasTagWithModuleVersion(TestUtils.TEST_GROUP_ID, "project-ear", "3.0"));
         assertThat(testProject.local, hasTagWithModuleVersion(TestUtils.TEST_GROUP_ID, "project-ejb", "3.0"));
@@ -37,9 +32,16 @@ public class EarProjectTest {
 
     @Test
     public void buildsAndInstallsEarWithCorrectVersions() throws Exception {
+        testProject.mvnReleaseComplete();
         final ZipFile zipFile = new ZipFile(testProject.getFile("project-ear", "target/project-ear-3.0.ear"));
         assertThat(zipFile.stream().map(ZipEntry::getName).collect(toList()), hasItems("project-ejb.jar",
                                                                                        "project.war",
                                                                                        "project-lib-2.0.jar"));
+    }
+
+    @Test
+    public void partialBuildWithoutPush() {
+        testProject.mvnReleasePrepare();
+        testProject.mvnInstall();
     }
 }